requirements.txt
fastapi
python-dotenv
python-multipart
pymongo
uvicorn
fuzzywuzzy
requests
python-jose[cryptography]
passlib[bcrypt]